Improve veech group and add origami list
Veech group:
We can now calculate the orbit of an origami class under the action
of SL(2,Z) via the `orb` package.

Origami list:
We can calculate a system of representatives of isomorphy classes of
origamis of given degree. We return canonical representatives.

We might need to change the action of SL(2,Z) to a right action.